refactor: move fn param utils into own util mod
